python-runner/requirements.txt
# Jupiter swap dependencies
requests>=2.31.0
solders>=0.21.0
base58>=2.1.0
python-dotenv>=1.0.0
